+++ src/Admin/AdminModule.php
@@ -10,6 +10,7 @@
 
 namespace Pronamic\WordPress\Pay\Admin;
 
+use Pronamic\WordPress\Money\Parser as MoneyParser;
 use Pronamic\WordPress\Pay\Plugin;
 
 /**
@@ -434,7 +435,7 @@ class AdminModule {
 				'tippy.js',
 				plugins_url( 'assets/tippy.js/tippy.all' . $min . '.js', $this->plugin->get_file() ),
 				array(),
-				'2.5.3',
+				'2.5.4',
 				true
 			);
 
@@ -477,14 +478,11 @@ class AdminModule {
 			$gateway = \Pronamic\WordPress\Pay\Plugin::get_gateway( $id );
 
 			if ( $gateway ) {
-				$amount = filter_input(
-					INPUT_POST, 'test_amount', FILTER_VALIDATE_FLOAT, array(
-						'flags'   => FILTER_FLAG_ALLOW_THOUSAND,
-						'options' => array(
-							'decimal' => pronamic_pay_get_decimal_separator(),
-						),
-					)
-				);
+				$string = filter_input( INPUT_POST, 'test_amount', FILTER_SANITIZE_STRING );
+
+				$money_parser = new MoneyParser();
+
+				$amount = $money_parser->parse( $string )->get_amount();
 
 				$data = new \Pronamic\WordPress\Pay\Payments\PaymentTestData( wp_get_current_user(), $amount );