+++ src/Admin/AdminSubscriptionPostType.php
@@ -10,6 +10,8 @@
 
 namespace Pronamic\WordPress\Pay\Admin;
 
+use Pronamic\WordPress\Money\Parser as MoneyParser;
+use Pronamic\WordPress\Pay\Core\Statuses;
 use Pronamic\WordPress\Pay\Plugin;
 use Pronamic\WordPress\Pay\Subscriptions\Subscription;
 use Pronamic\WordPress\Pay\Subscriptions\SubscriptionPostType;
@@ -18,9 +20,9 @@ use WP_Post;
 /**
  * WordPress admin subscription post type
  *
- * @author Reüel van der Steege
- * @version 1.0.0
- * @since unreleased
+ * @author  Reüel van der Steege
+ * @version 2.0.3
+ * @since   1.0.0
  */
 class AdminSubscriptionPostType {
 	/**
@@ -53,6 +55,8 @@ class AdminSubscriptionPostType {
 		add_action( 'manage_' . self::POST_TYPE . '_posts_custom_column', array( $this, 'custom_columns' ), 10, 2 );
 
 		add_action( 'add_meta_boxes', array( $this, 'add_meta_boxes' ) );
+
+		add_action( 'save_post_' . self::POST_TYPE, array( $this, 'save_post' ) );
 
 		add_filter( 'post_row_actions', array( $this, 'post_row_actions' ), 10, 2 );
 
@@ -350,17 +354,17 @@ class AdminSubscriptionPostType {
 	private function translate_post_status_to_meta_status( $post_status ) {
 		switch ( $post_status ) {
 			case 'subscr_pending':
-				return \Pronamic\WordPress\Pay\Core\Statuses::OPEN;
+				return Statuses::OPEN;
 			case 'subscr_cancelled':
-				return \Pronamic\WordPress\Pay\Core\Statuses::CANCELLED;
+				return Statuses::CANCELLED;
 			case 'subscr_expired':
-				return \Pronamic\WordPress\Pay\Core\Statuses::EXPIRED;
+				return Statuses::EXPIRED;
 			case 'subscr_failed':
-				return \Pronamic\WordPress\Pay\Core\Statuses::FAILURE;
+				return Statuses::FAILURE;
 			case 'subscr_active':
-				return \Pronamic\WordPress\Pay\Core\Statuses::ACTIVE;
+				return Statuses::ACTIVE;
 			case 'subscr_completed':
-				return \Pronamic\WordPress\Pay\Core\Statuses::COMPLETED;
+				return Statuses::COMPLETED;
 		}
 	}
 
@@ -390,5 +394,75 @@ class AdminSubscriptionPostType {
 		$subscription->set_status( $new_status_meta );
 
 		pronamic_pay_plugin()->subscriptions_data_store->update_meta_status( $subscription );
+	}
+
+	/**
+	 * When the post is saved, saves our custom data.
+	 *
+	 * @param int $post_id The ID of the post being saved.
+	 */
+	public function save_post( $post_id ) {
+		if ( ! filter_has_var( INPUT_POST, 'pronamic_subscription_update_nonce' ) ) {
+			return;
+		}
+
+		if ( ! check_admin_referer( 'pronamic_subscription_update', 'pronamic_subscription_update_nonce' ) ) {
+			return;
+		}
+
+		if ( 'pronamic_pay_subscr' !== get_post_type( $post_id ) ) {
+			return;
+		}
+
+		// If this is an autosave, our form has not been submitted, so we don't want to do anything.
+		if ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) {
+			return;
+		}
+
+		// Check the user's permissions.
+		if ( ! current_user_can( 'edit_post', $post_id ) ) {
+			return;
+		}
+
+		/* OK, its safe for us to save the data now. */
+		$definition = array(
+			'amount' => 'sanitize_text_field',
+		);
+
+		$subscription = new Subscription( $post_id );
+
+		$update_meta = array();
+
+		foreach ( $definition as $meta => $function ) {
+			$meta_key   = pronamic_pay_plugin()->subscriptions_data_store->meta_key_prefix . $meta;
+			$meta_value = null;
+
+			if ( 'sanitize_text_field' === $function ) {
+				if ( isset( $_POST[ $meta_key ] ) ) { // WPCS: input var OK.
+					$meta_value = sanitize_text_field( wp_unslash( $_POST[ $meta_key ] ) ); // WPCS: input var OK.
+				}
+			} else {
+				$filter  = $function;
+				$options = null;
+
+				if ( is_array( $function ) && isset( $function['filter'] ) ) {
+					$filter  = $function['filter'];
+					$options = $function;
+				}
+
+				$meta_value = filter_input( INPUT_POST, $meta_key, $filter, $options );
+			}
+
+			// Convert user input amount to float.
+			if ( 'amount' === $meta ) {
+				$money_parser = new MoneyParser();
+
+				$meta_value = $money_parser->parse( $meta_value )->get_amount();
+			}
+
+			$update_meta[ $meta ] = $meta_value;
+		}
+
+		$subscription->update_meta( $update_meta );
 	}
 }